requirements.txt
asgiref==3.8.1
async-timeout==5.0.1

certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.6.0
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
daphne==4.1.2
ecdsa==0.19.1
google-auth==2.38.0
idna==3.10
imagekitio
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.4.0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing-extensions==4.13.0
urllib3==2.2.3
celery==5.5.3
channels==4.2.2
channels_redis
python-decouple
